requirements.txt
requests==2.31.0
mysql-connector-python==8.1.0
flask==2.3.2
beautifulsoup4==4.12.2
lxml==4.9.3
schedule==1.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
openai==0.28.1
